Added a class that uses PDFPig
