+++ src/Expelibrum/Expelibrum.Services/Expelibrum.Services.csproj
@@ -6,6 +6,7 @@
 
   <ItemGroup>
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+    <PackageReference Include="PdfPig" Version="0.1.4" />
     <PackageReference Include="PdfSharpCore" Version="1.2.11" />
   </ItemGroup>
